rename: addteam dir team으로 수정
